+++ src/Eloquent/AbstractManyRelation.php
@@ -36,24 +36,9 @@ abstract class AbstractManyRelation extends AbstractRelation implements HasManyA
      */
     public function query($record, EncodingParametersInterface $parameters)
     {
-        /** If we do not need to pass to the inverse adapter, we can just return the whole relationship. */
-        if (!$this->requiresInverseAdapter($record, $parameters)) {
-            return $record->{$this->key};
-        }
+        $relation = $this->getRelation($record, $this->key);
 
-        $relation = $this->getRelation($record);
-
-        return $this->adapterFor($relation)->queryRelation($relation, $parameters);
-    }
-
-    /**
-     * @param Model $record
-     * @param EncodingParametersInterface $parameters
-     * @return mixed
-     */
-    public function relationship($record, EncodingParametersInterface $parameters)
-    {
-        return $this->query($record, $parameters);
+        return $this->adapterFor($relation)->queryToMany($relation, $parameters);
     }
 
 }